Don't transform classes with static props
This transform is beginning to look worse.. I did some further reading
but I don't think that converting to ES2015 classes is feasable as well
(http://benmccormick.org/2015/04/07/es6-classes-and-backbone-js/). So
instead I just figured we should tie this down more and more. To make
sure that it absolutely doesn't convert something that it shouldn't.